only video owner allow to edit delete video

+++ controllers/videoController.js
@@ -62,6 +62,9 @@ export const editVideo = async (req, res) => {
     if(!video){
         return res.status(404).render("404", {pageTitle: "video not found."})
     }
+    if(String(video.owner) !== String(req.session.user._id)){
+        return res.status(403).redirect("/");
+    }
     return res.render("editVideo", { pageTitle: `Editing ${video.title}`, video});
 }
 export const postEdit = async (req, res) => {
@@ -70,6 +73,9 @@ export const postEdit = async (req, res) => {
     const video = await Video.exists({ _id: id });
     if(!video){
         return res.render("404", {pageTitle: "video not found."})
+    }
+    if(String(video.owner) !== String(req.session.user._id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, 
@@ -81,6 +87,13 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "video not found."})
+    }
+    if(String(video.owner) !== String(req.session.user._id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
